+++ libkbfs/bare_root_metadata_v2_test.go
@@ -7,11 +7,15 @@ package libkbfs
 import (
 	"context"
 	"fmt"
+	"sort"
 	"testing"
 
+	"github.com/keybase/client/go/externals"
 	"github.com/keybase/client/go/protocol/keybase1"
+	"github.com/keybase/go-codec/codec"
 	"github.com/keybase/kbfs/kbfscodec"
 	"github.com/keybase/kbfs/kbfscrypto"
+	"github.com/keybase/kbfs/kbfshash"
 	"github.com/keybase/kbfs/tlf"
 	"github.com/stretchr/testify/require"
 )
@@ -47,6 +51,268 @@ func TestBareRootMetadataVersionV2(t *testing.T) {
 
 	rmd.SetUnresolvedWriters(nil)
 	require.Equal(t, PreExtraMetadataVer, rmd.Version())
+}
+
+// Test that old encoded WriterMetadataV2 objects (i.e., without any
+// extra fields) can be deserialized and serialized to the same form,
+// which is important for BareRootMetadataV2.IsValidAndSigned().
+func testWriterMetadataV2UnchangedEncoding(t *testing.T, ver MetadataVer) {
+	encodedWm := []byte{
+		0x89, 0xa3, 0x42, 0x49, 0x44, 0xc4, 0x10, 0x0,
+		0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
+		0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0xa9,
+		0x44, 0x69, 0x73, 0x6b, 0x55, 0x73, 0x61, 0x67,
+		0x65, 0x64, 0xa2, 0x49, 0x44, 0xc4, 0x10, 0x1,
+		0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
+		0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x16, 0xb3,
+		0x4c, 0x61, 0x73, 0x74, 0x4d, 0x6f, 0x64, 0x69,
+		0x66, 0x79, 0x69, 0x6e, 0x67, 0x57, 0x72, 0x69,
+		0x74, 0x65, 0x72, 0xa4, 0x75, 0x69, 0x64, 0x31,
+		0xa8, 0x52, 0x65, 0x66, 0x42, 0x79, 0x74, 0x65,
+		0x73, 0x63, 0xaa, 0x55, 0x6e, 0x72, 0x65, 0x66,
+		0x42, 0x79, 0x74, 0x65, 0x73, 0x65, 0xa6, 0x57,
+		0x46, 0x6c, 0x61, 0x67, 0x73, 0xa, 0xa7, 0x57,
+		0x72, 0x69, 0x74, 0x65, 0x72, 0x73, 0x92, 0xa4,
+		0x75, 0x69, 0x64, 0x31, 0xa4, 0x75, 0x69, 0x64,
+		0x32, 0xa4, 0x64, 0x61, 0x74, 0x61, 0xc4, 0x2,
+		0xa, 0xb,
+	}
+
+	expectedWm := WriterMetadataV2{
+		SerializedPrivateMetadata: []byte{0xa, 0xb},
+		LastModifyingWriter:       "uid1",
+		Writers:                   []keybase1.UID{"uid1", "uid2"},
+		ID:                        tlf.FakeID(1, false),
+		BID:                       NullBranchID,
+		WFlags:                    0xa,
+		DiskUsage:                 100,
+		RefBytes:                  99,
+		UnrefBytes:                101,
+	}
+
+	c := kbfscodec.NewMsgpack()
+
+	var wm WriterMetadataV2
+	err := c.Decode(encodedWm, &wm)
+	require.NoError(t, err)
+
+	require.Equal(t, expectedWm, wm)
+
+	buf, err := c.Encode(wm)
+	require.NoError(t, err)
+	require.Equal(t, encodedWm, buf)
+}
+
+func TestWriterMetadataV2UnchangedEncoding(t *testing.T) {
+	runTestOverMetadataVers(t, testWriterMetadataV2UnchangedEncoding)
+}
+
+// Test that WriterMetadataV2 has only a fixed (frozen) set of fields.
+func TestWriterMetadataV2EncodedFields(t *testing.T) {
+	sa1, _ := externals.NormalizeSocialAssertion("uid1@twitter")
+	sa2, _ := externals.NormalizeSocialAssertion("uid2@twitter")
+	// Usually exactly one of Writers/WKeys is filled in, but we
+	// fill in both here for testing.
+	wm := WriterMetadataV2{
+		ID:      tlf.FakeID(0xa, false),
+		Writers: []keybase1.UID{"uid1", "uid2"},
+		WKeys:   TLFWriterKeyGenerationsV2{{}},
+		Extra: WriterMetadataExtra{
+			UnresolvedWriters: []keybase1.SocialAssertion{sa1, sa2},
+		},
+	}
+
+	c := kbfscodec.NewMsgpack()
+
+	buf, err := c.Encode(wm)
+	require.NoError(t, err)
+
+	var m map[string]interface{}
+	err = c.Decode(buf, &m)
+	require.NoError(t, err)
+
+	expectedFields := []string{
+		"BID",
+		"DiskUsage",
+		"ID",
+		"LastModifyingWriter",
+		"RefBytes",
+		"UnrefBytes",
+		"WFlags",
+		"WKeys",
+		"Writers",
+		"data",
+		"x",
+	}
+
+	var fields []string
+	for field := range m {
+		fields = append(fields, field)
+	}
+	sort.Strings(fields)
+	require.Equal(t, expectedFields, fields)
+}
+
+type writerMetadataExtraFuture struct {
+	WriterMetadataExtra
+	kbfscodec.Extra
+}
+
+func (wmef writerMetadataExtraFuture) toCurrent() WriterMetadataExtra {
+	return wmef.WriterMetadataExtra
+}
+
+type tlfWriterKeyGenerationsV2Future []*tlfWriterKeyBundleV2Future
+
+func (wkgf tlfWriterKeyGenerationsV2Future) toCurrent() TLFWriterKeyGenerationsV2 {
+	wkg := make(TLFWriterKeyGenerationsV2, len(wkgf))
+	for i, wkbf := range wkgf {
+		wkb := wkbf.toCurrent()
+		wkg[i] = wkb
+	}
+	return wkg
+}
+
+type writerMetadataV2Future struct {
+	WriterMetadataV2
+	// Override WriterMetadata.WKeys.
+	WKeys tlfWriterKeyGenerationsV2Future
+	// Override WriterMetadata.Extra.
+	Extra writerMetadataExtraFuture `codec:"x,omitempty,omitemptycheckstruct"`
+}
+
+func (wmf writerMetadataV2Future) toCurrent() WriterMetadataV2 {
+	wm := wmf.WriterMetadataV2
+	wm.WKeys = wmf.WKeys.toCurrent()
+	wm.Extra = wmf.Extra.toCurrent()
+	return wm
+}
+
+func (wmf writerMetadataV2Future) ToCurrentStruct() kbfscodec.CurrentStruct {
+	return wmf.toCurrent()
+}
+
+func makeFakeWriterMetadataV2Future(t *testing.T) writerMetadataV2Future {
+	wmd := WriterMetadataV2{
+		// This needs to be list format so it fails to compile if new fields
+		// are added, effectively checking at compile time whether new fields
+		// have been added
+		[]byte{0xa, 0xb},
+		"uid1",
+		[]keybase1.UID{"uid1", "uid2"},
+		nil,
+		tlf.FakeID(1, false),
+		NullBranchID,
+		0xa,
+		100,
+		99,
+		101,
+		WriterMetadataExtra{},
+	}
+	wkb := makeFakeTLFWriterKeyBundleV2Future(t)
+	sa, _ := externals.NormalizeSocialAssertion("foo@twitter")
+	return writerMetadataV2Future{
+		wmd,
+		tlfWriterKeyGenerationsV2Future{&wkb},
+		writerMetadataExtraFuture{
+			WriterMetadataExtra{
+				// This needs to be list format so it fails to compile if new
+				// fields are added, effectively checking at compile time
+				// whether new fields have been added
+				[]keybase1.SocialAssertion{sa},
+				codec.UnknownFieldSetHandler{},
+			},
+			kbfscodec.MakeExtraOrBust("WriterMetadata", t),
+		},
+	}
+}
+
+func TestWriterMetadataV2UnknownFields(t *testing.T) {
+	testStructUnknownFields(t, makeFakeWriterMetadataV2Future(t))
+}
+
+type tlfReaderKeyGenerationsV2Future []*tlfReaderKeyBundleV2Future
+
+func (rkgf tlfReaderKeyGenerationsV2Future) toCurrent() TLFReaderKeyGenerationsV2 {
+	rkg := make(TLFReaderKeyGenerationsV2, len(rkgf))
+	for i, rkbf := range rkgf {
+		rkb := rkbf.toCurrent()
+		rkg[i] = rkb
+	}
+	return rkg
+}
+
+// rootMetadataWrapper exists only to add extra depth to fields
+// in RootMetadata, so that they may be overridden in
+// bareRootMetadataV2Future.
+type bareRootMetadataWrapper struct {
+	BareRootMetadataV2
+}
+
+type bareRootMetadataV2Future struct {
+	// Override BareRootMetadata.WriterMetadata. Put it first to work
+	// around a bug in codec's field lookup code.
+	//
+	// TODO: Report and fix this bug upstream.
+	writerMetadataV2Future
+
+	bareRootMetadataWrapper
+	// Override BareRootMetadata.RKeys.
+	RKeys tlfReaderKeyGenerationsV2Future `codec:",omitempty"`
+	kbfscodec.Extra
+}
+
+func (brmf *bareRootMetadataV2Future) toCurrent() BareRootMetadata {
+	rm := brmf.bareRootMetadataWrapper.BareRootMetadataV2
+	rm.WriterMetadataV2 = WriterMetadataV2(brmf.writerMetadataV2Future.toCurrent())
+	rm.RKeys = brmf.RKeys.toCurrent()
+	return &rm
+}
+
+func (brmf *bareRootMetadataV2Future) ToCurrentStruct() kbfscodec.CurrentStruct {
+	return brmf.toCurrent()
+}
+
+func makeFakeBareRootMetadataV2Future(t *testing.T) *bareRootMetadataV2Future {
+	wmf := makeFakeWriterMetadataV2Future(t)
+	rkb := makeFakeTLFReaderKeyBundleV2Future(t)
+	h, err := kbfshash.DefaultHash([]byte("fake buf"))
+	require.NoError(t, err)
+	sa, _ := externals.NormalizeSocialAssertion("bar@github")
+	rmf := bareRootMetadataV2Future{
+		wmf,
+		bareRootMetadataWrapper{
+			BareRootMetadataV2{
+				// This needs to be list format so it
+				// fails to compile if new fields are
+				// added, effectively checking at
+				// compile time whether new fields
+				// have been added
+				WriterMetadataV2{},
+				kbfscrypto.SignatureInfo{
+					Version:      100,
+					Signature:    []byte{0xc},
+					VerifyingKey: kbfscrypto.MakeFakeVerifyingKeyOrBust("fake kid"),
+				},
+				"uid1",
+				0xb,
+				5,
+				MdID{h},
+				nil,
+				[]keybase1.SocialAssertion{sa},
+				nil,
+				nil,
+				codec.UnknownFieldSetHandler{},
+			},
+		},
+		[]*tlfReaderKeyBundleV2Future{&rkb},
+		kbfscodec.MakeExtraOrBust("BareRootMetadata", t),
+	}
+	return &rmf
+}
+
+func TestBareRootMetadataV2UnknownFields(t *testing.T) {
+	testStructUnknownFields(t, makeFakeBareRootMetadataV2Future(t))
 }
 
 func TestIsValidRekeyRequestBasicV2(t *testing.T) {
